pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dtSpark"
dynamic = ["version"]
description = "Secure Personal AI Research Kit - Multi-provider LLM CLI/Web interface with MCP tool integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Matthew Westwood-Hill", email = "matthew@digital-thought.org"}
]
keywords = [
    "llm",
    "ai",
    "chatbot",
    "aws",
    "bedrock",
    "anthropic",
    "claude",
    "ollama",
    "mcp",
    "model-context-protocol",
    "cli",
    "web",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.6.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "mcp>=0.9.0",
    "pyyaml>=6.0",
    "dtPyAppFramework>=4.0.3",
    "tiktoken>=0.5.0",
    "ollama>=0.2.0",
    "cryptography>=41.0.0",
    "anthropic>=0.18.0",
    "APScheduler>=3.10.0",
    "markdown>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]
mysql = ["mysql-connector-python>=8.0.0"]
postgresql = ["psycopg2-binary>=2.9.0"]
mssql = ["pyodbc>=4.0.0"]
all-databases = [
    "mysql-connector-python>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "pyodbc>=4.0.0",
]

[project.scripts]
spark = "dtSpark.launch:main"
dtSpark = "dtSpark.launch:main"

[project.urls]
Homepage = "https://github.com/digital-thought/dtSpark"
Documentation = "https://github.com/digital-thought/dtSpark#readme"
Repository = "https://github.com/digital-thought/dtSpark"
Issues = "https://github.com/digital-thought/dtSpark/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {file = "src/dtSpark/_version.txt"}

[tool.setuptools.package-data]
dtSpark = [
    "_*.txt",
    "_*.yaml",
    "resources/*.yaml",
    "resources/*.template",
    "web/templates/*.html",
    "web/static/css/*.css",
    "web/static/js/*.js",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
